+++ src/DateInput/styles.ts
@@ -7,7 +7,7 @@ import {
   Label,
   Text,
 } from 'react-aria-components';
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 import { defaultTheme } from '../defaultTheme';
 
 export const StyledDateField = styled(DateField)`