+++ src/Slider/styles.ts
@@ -5,7 +5,7 @@ import {
   SliderTrack,
   SliderThumb,
 } from 'react-aria-components';
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 import { defaultTheme } from '../defaultTheme';
 
 export const StyledSlider = styled(Slider)`