+++ src/Switch/styles.tsx
@@ -1,5 +1,5 @@
 import { Switch } from 'react-aria-components';
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 import { defaultTheme } from '../defaultTheme';
 
 export const Indicator = styled.div`